Split out the Hover request into its own file
The real reason for this? To remove CPP from DataTypesJSON, and thus to
remove a preprocessor error/warning caused by one of the TypeScript
documentation comments containing a /**
